// src/models/investorTrades.ts
export interface InvestorTrades {
    id_trade_type: string;
    capital: string;
    cash_rounded_capital: string;
    cash_rounded_settlement_amount: string;
    cash_rounded_receivable_payable_amount: string;
    purchased_perf_fee_factor: string;
    cost_of_inv_red_proceeds: string;
    billing_code: string;
    id_trade: string;
    trade_type_name: string;
    trade_sub_type_name: string;
    units: string;
    sign: string;
    total_fees: string;
    valuation_date: string;
    trade_date: string;
    order_trade_date: string;
    valuation: string;
    gav_pre_fees: string;
    ask: string;
    bid: string;
    total_perf_fee_factor: string;
    valuation_precision: string;
    units_precision: string;
    units_description: string;
    trade_settlement_amount: string;
    id_order: string;
    method: string;
    class_currency: string;
    trade_type_order: string;
    trade_estimate: string;
    admin_client_mantra_id: string;
    portfolio_mantra_id: string;
    legal_entity_mantra_id: string;
    legal_entity_description: string;
    class_mantra_id: string;
    class_description: string;
    sub_class_mantra_id: string;
    sub_class_description: string;
    nominee_mantra_id: string;
    investor_mantra_id: number;
    investor_description: string;
    investor_name: string;
    units_on_int_reports: string;
}